Refactor: split tools into focused modules
Split inline-tools.ts (1381→578 lines) and browser-tools.ts (1088→472 lines)
into focused modules:

- meeting-tools.ts: summon, dismiss, joinZoom, joinGmeet, lookupMeetingId, callContact
- recording-tools.ts: scrollAndDescribe, screenRecord, video playback tools,
  recording lifecycle (resetDemoState, setupRecordingHooks, etc.)
- recording-state.ts: shared demoStateRef to avoid circular imports

inline-tools.ts and browser-tools.ts re-export from the new modules so all
existing imports continue to work. No behavior changes.
